+++ shoop/admin/modules/orders/views/__init__.py
@@ -14,11 +14,12 @@ from .edit import OrderEditView, PurchaseOrderEditView
 from .list import OrderListView, PurchaseOrderListView
 from .log import NewLogEntryView
 from .payment import OrderCreatePaymentView
+from .refund import OrderCreateFullRefundView, OrderCreateRefundView
 from .shipment import OrderCreateShipmentView
 
 __all__ = [
     "NewLogEntryView", "OrderDetailView", "OrderEditView", "OrderListView",
-    "OrderCreatePaymentView", "OrderCreateShipmentView", "OrderSetStatusView",
-    "PurchaseOrderEditView", "PurchaseOrderDetailView", "PurchaseOrderListView",
-    "PurchaseOrderSetArrivedView"
+    "OrderCreatePaymentView", "OrderCreateFullRefundView", "OrderCreateRefundView",
+    "OrderCreateShipmentView", "OrderSetStatusView", "PurchaseOrderEditView",
+    "PurchaseOrderDetailView", "PurchaseOrderListView", "PurchaseOrderSetArrivedView",
 ]